refactor: change package structure make package names lowercased
