--- app/src/main/java/com/iitp/njack/iitp_connect/Utils/GeneralUtils.java
@@ -1,12 +1,0 @@
-package com.iitp.njack.iitp_connect.Utils;
-
-/**
- * Created by srv_twry on 19/8/17.
- * This class contains the general utilities for the project.
- * List of utilites:
- *
- */
-
-public class GeneralUtils {
-
-}
