+++ app/src/main/java/com/iitp/njack/iitp_connect/data/network/Client.java
@@ -1,4 +1,4 @@
-package com.iitp.njack.iitp_connect.Retrofit;
+package com.iitp.njack.iitp_connect.data.network;
 
 import okhttp3.ResponseBody;
 import retrofit2.Call;
