+++ app/src/main/res/layout/activity_coding_calendar_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.iitp.njack.iitp_connect.CodingCalendar.Activities.CodingCalendarHomeActivity">
+    tools:context=".core.codingcalendar.views.CodingCalendarHomeActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/rv_coding_calendar"
