+++ app/src/main/res/layout/activity_contest_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.iitp.njack.iitp_connect.CodingCalendar.Activities.ContestDetailsActivity">
+    tools:context=".core.codingcalendar.views.ContestDetailsActivity">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
