+++ app/src/main/res/layout/app_bar_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.iitp.njack.iitp_connect.HomeActivity">
+    tools:context="com.iitp.njack.iitp_connect.core.home.HomeActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
