+++ app/src/main/res/layout/content_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.iitp.njack.iitp_connect.HomeActivity"
+    tools:context="com.iitp.njack.iitp_connect.core.home.HomeActivity"
     tools:showIn="@layout/app_bar_home">
 
     <TextView
